% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakeoff-package.R
\docType{package}
\name{bakeoff-package}
\alias{bakeoff}
\alias{bakeoff-package}
\title{bakeoff: Data from The Great British Bake-Off}
\description{
Data about the episodes, bakers, series, and spin-off series for The Great British Bake-Off, from Wikipedia.
}
\author{
\strong{Maintainer}: Alison Hill \email{apresstats@gmail}

Authors:
\itemize{
  \item Chester Ismay \email{chester.ismay@gmail}
}

}
\keyword{internal}
