% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{bakeoff_color_data}
\alias{bakeoff_color_data}
\title{Define hex codes for named bakeoff colors}
\format{
An object of class \code{character} of length 63.
}
\usage{
bakeoff_color_data
}
\description{
Define hex codes for named bakeoff colors
}
\keyword{datasets}
