% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{bakeoff_colors}
\alias{bakeoff_colors}
\title{Extract named bakeoff colors as hex codes}
\usage{
bakeoff_colors(...)
}
\arguments{
\item{...}{Character names of bakeoff colors.
If none specified, returns all
If any specified, can be listed by names in quotes (no combine required)}
}
\value{
A named character vector of hex colors.
}
\description{
Extract named bakeoff colors as hex codes
}
\examples{
bakeoff_colors()
bakeoff_colors("riptide")
bakeoff_colors("baltic", "yellow")
names(bakeoff_colors())

library(scales)
show_col(bakeoff_colors(), label = FALSE)
}
