% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakeoff-colors.R
\name{bakeoff_cols}
\alias{bakeoff_cols}
\title{Function to extract bakeoff colors as hex codes}
\usage{
bakeoff_cols(...)
}
\arguments{
\item{...}{Character names of bakeoff_colors
If none specified, returns all
If any specified, can be listed by names in quotes (no combine required)}
}
\description{
Function to extract bakeoff colors as hex codes
}
