% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{bakeoff_generate_pal}
\alias{bakeoff_generate_pal}
\title{Return function that creates a bakeoff color palette
Mainly useful for creating ggplot2 scales}
\usage{
bakeoff_generate_pal(palette, direction)
}
\arguments{
\item{palette}{Character name of palette in bakeoff_palettes}

\item{direction}{Either \code{1} or \code{-1}. If \code{-1}, the palette will be reversed.}
}
\description{
Return function that creates a bakeoff color palette
Mainly useful for creating ggplot2 scales
}
