% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakeoff-colors.R
\name{bakeoff_pal}
\alias{bakeoff_pal}
\title{Return function to interpolate a bakeoff color palette}
\usage{
bakeoff_pal(palette = "main", reverse = FALSE)
}
\arguments{
\item{palette}{Character name of palette in bakeoff_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a bakeoff color palette
}
