% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{bakeoff_palette_names}
\alias{bakeoff_palette_names}
\title{Print names of all \strong{bakeoff} color palettes}
\usage{
bakeoff_palette_names()
}
\description{
Print names of all \strong{bakeoff} color palettes
}
\examples{
bakeoff_palette_names()

}
