% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{bakeoff_palettes}
\alias{bakeoff_palettes}
\title{Define bakeoff color palettes from named colors}
\format{
An object of class \code{list} of length 3.
}
\usage{
bakeoff_palettes
}
\description{
Define bakeoff color palettes from named colors
}
\keyword{datasets}
