% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakers.R
\docType{data}
\name{bakers}
\alias{bakers}
\title{Each baker's name, age, occupation, and hometown}
\format{
A data frame with 120 rows representing individual bakers and 8 variables:
\describe{
\item{series}{an integer denoting UK series (1-10)}
\item{baker_full}{a character string giving full name}
\item{baker}{a character string giving given or nickname}
\item{age}{an integer denoting age in years at first episode appeared}
\item{occupation}{a character string giving occupation}
\item{hometown}{a character string giving hometown}
\item{baker_last}{a character string giving family name}
\item{baker_first}{a character string giving given name}
}
}
\source{
See \url{https://en.wikipedia.org/wiki/The_Great_British_Bake_Off_(series_1)#The_Bakers}, for example, for series 1 bakers.
}
\usage{
bakers
}
\description{
Bakers across all 10 GBBO series
}
\keyword{datasets}
