% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/challenge_results.R
\docType{data}
\name{challenge_results}
\alias{challenge_results}
\title{Each baker's signature, technical, and showstopper challenges, plus their fate, by episode}
\format{
A data frame with 886 rows representing individual bakers per episode and 7 variables:
\describe{
\item{series}{an integer denoting UK series (1-8)}
\item{episode}{an integer denoting episode number within a series}
\item{baker}{a character string giving given or nickname}
\item{result}{a character string denoting if the baker was \code{IN}, \code{OUT}, \verb{STAR BAKER}, or \code{SICK} for a given episode. For finale episodes, values are either \code{WINNER} or \code{RUNNER-UP}. If \code{NA}, baker did not appear in episode.}
\item{signature}{a character string containing the bake for the signature challenge for that baker/episode. If \code{NA}, baker did not appear in episode.}
\item{technical}{an integer denoting the rank on the technical challenge for that baker/episode; rank = 1 is winner of the technical challenge. If \code{NA}, baker did not appear in episode.}
\item{showstopper}{a character string containing the bake for the showstopper challenge for that baker/episode. If \code{NA}, baker did not appear in episode.}
}
}
\source{
See \url{https://en.wikipedia.org/wiki/The_Great_British_Bake_Off#Format}
}
\usage{
challenge_results
}
\description{
Details about the three challenges (signature, technical, and showstopper) for each baker/episode across all GBBO series, and who was eliminated, who won star baker, or if finale, who won and who was runner up. This joined tibble is made by joining the \code{\link{challenges}} and \code{\link{results}} simple tibbles by \code{series}, \code{episode}, and \code{baker}.
}
\keyword{datasets}
