% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{data}
\name{results}
\alias{results}
\title{Each baker's fate, by episode}
\format{A data frame with 886 rows representing individual bakers per episode and 4 variables:
\describe{
\item{series}{an integer denoting UK series (1-8)}
\item{episode}{an integer denoting episode number within a series}
\item{baker}{a character string giving given or nickname only (note: see \code{bakers} for full baker names)}
\item{result}{a character string denoting if the baker was \code{IN}, \code{OUT}, \code{STAR BAKER}, or \code{SICK} for a given episode. For finale episodes, values are either \code{WINNER} or \code{RUNNER-UP}. If \code{NA}, baker did not appear in episode.}
}}
\source{
See \url{https://en.wikipedia.org/wiki/The_Great_British_Bake_Off_(series_1)#Results_summary}, for example, for series 1 results summary.
}
\usage{
results
}
\description{
Results for all bakers per episode across all GBBO series
}
\keyword{datasets}
