% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scales.R
\name{scale_bakeoff}
\alias{scale_bakeoff}
\alias{scale_color_bakeoff}
\alias{scale_fill_bakeoff}
\title{Color scale constructor for bakeoff colors}
\usage{
scale_color_bakeoff(
  palette = "showstopper",
  discrete = TRUE,
  direction = 1,
  ...
)

scale_fill_bakeoff(
  palette = "showstopper",
  discrete = TRUE,
  direction = 1,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in bakeoff_palette}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{direction}{Either \code{1} or \code{-1}. If \code{-1}, the palette will be reversed.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for bakeoff colors

Fill scale constructor for bakeoff colors
}
\examples{

if (require('ggplot2')) {

  ggplot(ratings, aes(x = episode, y = viewers_7day, group = series, color = series)) +
    geom_line(lwd = 3) +
    theme_minimal() +
    scale_color_bakeoff("finale", guide = FALSE)
}

if (require('ggplot2')) {

  ggplot(ratings, aes(x = episode_count, y = viewers_7day, fill = series)) +
    geom_col() +
    theme_minimal() +
    scale_fill_bakeoff("finale", guide = FALSE)
}
}
\concept{colour scales}
