% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spice_test_wide.R
\docType{data}
\name{spice_test_wide}
\alias{spice_test_wide}
\title{Spice Test}
\format{
A data frame with 4 rows representing baker results from the
spice-based knowledge test and 7 variables:
\describe{
\item{baker}{A character string with a given name or nickname.}
\item{guess_1,guess_2,guess_3}{The different guesses (in order) for what
the mystery spice was.}
\item{correct_1,correct_2,correct_3}{An integer whether the guess was
correct (\code{1}) or wrong (\code{0}).}
}
}
\usage{
spice_test_wide
}
\description{
Results of a knowledge test from Junior Bake Off (Series 2, Episode 14).
}
\examples{
if (require('tibble')) {
  spice_test_wide
 }
head(spice_test_wide)
}
\keyword{datasets}
