% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakeoff-package.R
\docType{package}
\name{bakeoff-package}
\alias{bakeoff}
\alias{bakeoff-package}
\title{bakeoff: Data from The Great British Bake-Off}
\description{
Data about the episodes, bakers, series, and spin-off series for The Great British Bake-Off, from Wikipedia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bakeoff.netlify.com}
  \item \url{https://github.com/apreshill/bakeoff}
  \item Report bugs at \url{https://github.com/apreshill/bakeoff/issues}
}

}
\author{
\strong{Maintainer}: Alison Hill \email{apresstats@gmail} (https://orcid.org/0000-0002-8082-1890)

Authors:
\itemize{
  \item Chester Ismay \email{chester.ismay@gmail}
}

}
\keyword{internal}
