% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakeoff-colors.R
\docType{data}
\name{bakeoff_colors}
\alias{bakeoff_colors}
\title{Define the colors}
\format{
An object of class \code{character} of length 25.
}
\usage{
bakeoff_colors
}
\description{
Define the colors
}
\keyword{datasets}
