% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\docType{data}
\name{seasons}
\alias{seasons}
\title{Seasons and airdates for the US version called 'The Great British Baking Show'}
\format{A data frame with 8 rows representing individual series and 11 variables:
\describe{
\item{series}{an integer denoting UK series (1-8)}
\item{episodes}{an integer denoting total number of episodes within series}
\item{us_season}{an integer denoting US season (1-5)}
\item{us_airdate}{a date denoting original airdate of episode in the US, according to \href{http://www.pbs.org/food/shows/great-british-baking-show/}{pbs.org}}
}}
\source{
US airdates manually recorded from \url{http://www.pbs.org/food/shows/great-british-baking-show/}
}
\usage{
seasons
}
\description{
US seasons and airdates for all GBBO series/episodes
}
\keyword{datasets}
